+++ tests/integration/test_spans.py
@@ -534,6 +534,11 @@ def test_span_ingestion(
                 key="sentry.exclusive_time_nano",
                 value=AnyValue(int_value=int(duration.total_seconds() * 1e9)),
             ),
+            # In order to test `category` sentry tag inference.
+            KeyValue(
+                key="ui.component_name",
+                value=AnyValue(string_value="MyComponent"),
+            ),
         ],
     )
     scope_spans = ScopeSpans(spans=[protobuf_span])
@@ -699,6 +704,7 @@ def test_span_ingestion(
             "data": {
                 "browser.name": "Python Requests",
                 "client.address": "127.0.0.1",
+                "ui.component_name": "MyComponent",
                 "user_agent.original": "python-requests/2.32.2",
             },
             "description": "my 3rd protobuf OTel span",
@@ -712,6 +718,7 @@ def test_span_ingestion(
             "retention_days": 90,
             "sentry_tags": {
                 "browser.name": "Python Requests",
+                "category": "ui",
                 "op": "default",
                 "status": "unknown",
             },